% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree_table.R
\name{agree_table}
\alias{agree_table}
\title{Calculate agreement between two nominal/ordinal variables}
\usage{
agree_table(x, y, cohen_k = c("uw", "wl", "wq"), conf.level = 0.95, ...)
}
\arguments{
\item{x}{first qualitative variable}

\item{y}{first second variable}

\item{cohen_k}{weight for cohen (uw = unweighted, 'wl = linearly weighted',}

\item{conf.level}{confidence level for estimate confidence interval}

\item{...}{further arguments passed to biv_quali}
}
\description{
Calculate agreement between two nominal/ordinal variables
}
