% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{fleiss_k}
\alias{fleiss_k}
\title{Fleiss' kappa}
\usage{
fleiss_k(
  x,
  nlevels = NULL,
  raters_par = list(test = "MC", B = 1000, alpha = 0.05),
  irr_par = list(exact = FALSE, detail = FALSE),
  gwet_par = list(weights = "unweighted", categ = NULL, conflev = 0.95, N = Inf)
)
}
\arguments{
\item{x}{a data.frame with one column per scale item/subject and
one row per rater. Judgement can be a factors or integers}

\item{nlevels}{number of levels of possible rating/judgement}

\item{raters_par}{arguments passed to raters::concordance}

\item{irr_par}{arguments passed to irr::kappam.fleiss}

\item{gwet_par}{arguments passed to irrCAC::gwet.ac1.dist}
}
\description{
Fleiss' kappa (Cohen's equivalent for multiple raters)
}
\examples{
## reproducing the example from
## https://en.wikipedia.org/wiki/Fleiss\%27_kappa
item1  <- rep(5, 14)
item2  <- c(rep(1,0), rep(2,2), rep(3,6), rep(4,4), rep(5,2))
item3  <- c(rep(1,0), rep(2,0), rep(3,3), rep(4,5), rep(5,6))
item4  <- c(rep(1,0), rep(2,3), rep(3,9), rep(4,2), rep(5,0))
item5  <- c(rep(1,2), rep(2,2), rep(3,8), rep(4,1), rep(5,1))
item6  <- c(rep(1,7), rep(2,7), rep(3,0), rep(4,0), rep(5,0))
item7  <- c(rep(1,3), rep(2,2), rep(3,6), rep(4,3), rep(5,0))
item8  <- c(rep(1,2), rep(2,5), rep(3,3), rep(4,2), rep(5,2))
item9  <- c(rep(1,6), rep(2,5), rep(3,2), rep(4,1), rep(5,0))
item10 <- c(rep(1,0), rep(2,2), rep(3,2), rep(4,3), rep(5,7))

df <-  data.frame(item1, item2, item3, item4, item5,
                  item6, item7, item8, item9, item10)
df
fleiss_k(x = df, nlevels = 5)

}
