% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_rho.R
\name{lin_diff}
\alias{lin_diff}
\title{computes difference between lin CCC (and confidence interval)
comparing x and y with a reference standard}
\usage{
lin_diff(x, y, refstd, conf = 0.95, R = 10000)
}
\arguments{
\item{x}{first measurement (classical)}

\item{y}{second measurement (experimental)}

\item{refstd}{reference standard for measurement}

\item{conf}{confidence level}

\item{R}{bootstrap replicates}
}
\description{
computes difference between lin CCC (and confidence interval)
comparing x and y with a reference standard
}
