% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_rho.R
\name{lin_plot}
\alias{lin_plot}
\title{Scatterplot with (0,0) to (1,1) line}
\usage{
lin_plot(x, y, stat = c("ci", "coef", "none"), ...)
}
\arguments{
\item{x}{generally the gold standard (on x axis)}

\item{y}{generally the testing measurement (on y axis)}

\item{stat}{which statistics to plot in a (topleft-placed) legend}
}
\description{
Scatterplot with (0,0) to (1,1) line
}
