% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_rho.R
\name{lin_rho}
\alias{lin_rho}
\title{Lin's Rho (or CCC)}
\usage{
lin_rho(x = NULL, y = NULL, ci = c("z-transform", "asymptotic"),
  conf.level = 0.95, rep.measure = FALSE, subjectid = NULL)
}
\arguments{
\item{x}{vector of first set of measurements}

\item{y}{vector of first set of measurements}

\item{ci}{method to be used for ci}

\item{conf.level}{magniture of the returned confidence interval}

\item{rep.measure}{if TRUE there are repeated observation across subject}

\item{subjectid}{factor providing id of the observer if rep.measure == TRUE}
}
\description{
Lin's agreement/concordance correlation coefficient (using
epiR::epi.ccc as workhorse)
}
