% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_rho.R
\name{occc}
\alias{occc}
\title{OCCC (overall concordance correlation coefficient)}
\usage{
occc(x, ...)
}
\arguments{
\item{x}{data.frame of measurement one column for rater}

\item{...}{further arguments passed to agRee::agree.ccc}
}
\description{
da citare barnhart2002occc
}
